Wwise Project Audio Architecture Foundation Created
I created an basic audio architecture that incorporates most of the basic sounds that we'll likely encounter in the early stages of development. Created numerous events for Player, Mobs, and Environment, as well as switches for material types that are hooked into events that play footfalls and object impacts. I also created RTPCs for player velocity, mob velocity, object velocity, and object mass. I create states that will select ambience tracks to play based on location, and an above/under-water state to account for that possibility. I also implemented an ambience audio file for a sunny day.